Made a page for the "About" menu item. Blank right now, for Camila to fill in with her work.
Tried to consolidate the widgets under the "Dashboard" menu item. Successfully did that, but now there's no content on the main body page. It must be not sure where to put the map/graph/table. Might not work out and we'll just have them separate in the sidebar menu...

Also changed the name, but need to fix the title width for the whole thing to fit.